Party Simulation
The PartySimulation program is an interactive multi-threaded simulation written in the
Java programming language. This simulation aims to model the dynamics of a real-world
party, where guests consume various foods and drinks while a waiter works to keep these
offerings continuously fresh. The multi-threading and synchronization mechanisms at the core
of the program are excellent examples for understanding complex interactions and resource
management in real-time systems. This simulation also serves as an indicator of how social
interactions and human behaviors at a party can be emulated through technology.